+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 CHANGELOG
 =========
 
-3.4.2a, 2015-05-11
+3.4.2a, 2015-05-13
 ------------------
+- applying or resetting message filter will scroll to last selection;
 - added support for copying selected list items to clipboard;
 - avoid caching messages on export to decrease memory overhead;
 - fixed a potential error with unexpected data in quoted messages;